+++ examples/bookstore/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod book;
+pub mod document;
 
 use driver::{
     client::{
@@ -6,7 +7,6 @@ use driver::{
         error::DatabaseClientError,
     },
     database::Database,
-    document::DocumentModel,
 };
 use std::env::current_exe;
 
@@ -19,27 +19,4 @@ pub fn create_db_client() -> DatabaseClient {
 
 pub fn get_bookstore_db(client: &DatabaseClient) -> Result<Database, DatabaseClientError> {
     Ok(client.get_database("bookdb")?)
-}
-
-/// Displays a document listing its data.
-pub fn display_document(document: &DocumentModel) {
-    println!("_id: {}", document.id());
-
-    for (key, value) in &document.data {
-        println!("{}: {}", key, value);
-    }
-}
-
-/// Displays a list of documents and their data.
-pub fn display_document_list(documents: &Vec<DocumentModel>) {
-    for document in documents {
-        println!("{{");
-        println!("  _id: {}", document.id);
-
-        for (key, value) in &document.data {
-            println!("  {}: {}", key, value);
-        }
-
-        println!("}}");
-    }
 }
